Removes sops.h
Moves all string operations into cite.c so code can be distributed as
a single file
Tidies up cite.c and removes unnecessary forward declarations
